requirements.txt
# Core dependencies for Iceberg TPC-DS Benchmark
pyspark==3.4.2
pandas>=1.5.0
numpy>=1.24.0
pyyaml>=6.0
click>=8.0.0

# Monitoring and metrics
psutil>=5.9.0

# Visualization (optional)
matplotlib>=3.6.0
seaborn>=0.12.0

# Kubernetes client (for controller)
kubernetes>=27.0.0
flask>=2.3.0

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter notebooks (optional)
jupyter>=1.0.0
jupyterlab>=4.0.0
